Creating a development branch.
This development branch will have on-going and incomplete work.

The implementation of a doubly linked list that is compatible with
STL-style iterators is almost completed. New tests need to be written
for this list.

The development of future trees and graphs not ready to be pushed to
master will be in this branch as well.