Restructured unit tests and sorting algorithms
The main change is related to the sorting algorithms. They have been
refactored to be iterator-based instead of container based so they will
should work with any container in practice, regardless of design
differences.

However, quicksort is currently broken. The current iterator-based
implementation of the partition function does not work and some work
needs to be done there.